GDB - Extract return value from target call
Ubi32 returns function values in D0 for values up to 4 bytes long.  For
values larger than 4 bytes, up to 8 bytes, values are returned in D0 and D1.
For little-endian targets, the high-order word is in d1 and the low-order
word is in d0.  For big-endian targets, this is reversed.

This patch extracts the values from D0 and D1, combines them if necessary,
and uses store_unsigned_integer to save the return value in the correct
byte order.

* gdb/ubi32-tdep.c(ubi32_push_dummy_call): Rewrite return value extraction.